requirements.txt
streamlit==1.24.0
sentence-transformers==2.2.2
chromadb==0.4.3
llama-cpp-python==0.1.77
PyMuPDF==1.22.5
numpy==1.24.3 